refactor: 유틸함수 생성 후 별도 파일로 분리
