+++ src/app/(main)/_component/CalendarUI.tsx
@@ -1,19 +1,11 @@
-import { getCalendarCellsInfo } from '@/shared/utils/getCalendarInfo'
-import { MonthIndex, CalendarUIProps } from '@/shared/types/type'
+import { MonthIndex, CalendarUIProps, CalendarCellInfo } from '@/shared/types/type'
 import { Grid, Box, Typography } from '@mui/material'
 import CalendarCell from './CalendarCell'
+import getTotalCalendarCellsInfo from '../_utils/getTotalCalendarCellsInfo'
 
 export default function CalendarUI({ selectedYear, selectedMonth }: CalendarUIProps) {
-  const calendarCellsInfo = getCalendarCellsInfo({ year: selectedYear, monthIndex: (selectedMonth - 1) as MonthIndex })
-  const calendarCellsInfoArray = Object.values(calendarCellsInfo).map((cellInfo) => ({
-    ...cellInfo,
-    isDummyCell: false,
-  }))
-  const dummyCellCount: number = calendarCellsInfoArray[0].dayIndex
-  const dummyCellsInfoArray = [...Array(dummyCellCount)].map((value, i) => ({ key: `0-${i}`, isDummyCell: true }))
-  // todo valid cell, dummy cell 구분
-  // todo getTotalCalendarCellsInfo 함수로 분리 하여 모듈화 하는게 좋은 것 같아 보임
-  const totalCellsInfoArray = [...dummyCellsInfoArray, ...calendarCellsInfoArray]
+  const monthIndex = (selectedMonth - 1) as MonthIndex
+  const totalCellsInfoArray = getTotalCalendarCellsInfo(selectedYear, monthIndex)
 
   const DAYS_HEADERS = [
     { day: '일', key: 'sunday' },
@@ -39,7 +31,7 @@ export default function CalendarUI({ selectedYear, selectedMonth }: CalendarUIPr
 
       {totalCellsInfoArray.map((cellInfo) => (
         <Grid item xs={1} key={cellInfo.key}>
-          {cellInfo.isDummyCell ? <CalendarCell /> : <CalendarCell cellInfo={cellInfo} />}
+          {cellInfo.isDummyCell ? <CalendarCell /> : <CalendarCell cellInfo={cellInfo as CalendarCellInfo} />}
         </Grid>
       ))}
     </Grid>
